misc: fuse: make type of tegra_get_chip_id() to consistent
For some Tegra SOCs, the function tegra_get_chip_id() returns u8
and for some other it returns u32.

Make the return type to be consistent and should return u8 data type.

bug 200260692

Reviewed-on: http://git-master/r/1268604
GVS: Gerrit_Virtual_Submit

+++ drivers/misc/tegra-fuse/tegra_fuse.c
@@ -448,7 +448,7 @@ u32 tegra_get_sku_id(void)
 	return tegra_chip_sku_id;
 }
 
-u32 tegra_get_chip_id(void)
+u8 tegra_get_chip_id(void)
 {
 	return tegra_chip_id;
 }
